Update Find-ServiceUser.ps1
In our corporate network we have WinRM open but frequently ICMP is not. Because Get-CimInstance already has error handling, I thikn it might make sense to not perform the ping check before running Get-CimInstance.

Also, the write-verbose 'querying wmi' part is a bit noisy in combination with the verbose output of get-ciminstance, and it's a bit inaccurate to say it's querying WMI when it's hitting WinRM with get-ciminstance. I think wmi implies you're using high ports.